src/agents/audio transcribe/requirements.txt
streamlit
openai-whisper
torch
numpy<=2.3.0
soundfile
av
openai-whisper
torch
numpy<=2.3.0
sounddevice
scipy
